joystick.py
joystick.py is a basic program I made to pull button state and axis position values from a standard HID joystick. Ideally this program is going to be run on a Raspberry Pi, which will then send this data to an Arduino, which will be "in charge" of sending commands to the stepper motors on the Pig Pen. See comments for more info.

I code, but I'm not a software engineer. Suggestions to improve the code are most welcome.